transfer
git-svn-id: svn+ssh://gitlab/srv/svn-repos/reference-annotation-builder/trunk@21 41b9b187-71fe-46d0-9756-c8471d252771

+++ src/screendata.cpp
@@ -154,7 +154,7 @@ void ScreenData::analyze(const std::string& assembly, unsigned readLength, const
 				auto ti = t + i;
 				long d = ti->chrom - chr;
 				if (d == 0)
-					d = ti->r.start - pos;
+					d = static_cast<long>(ti->r.start) - pos;
 				if (d >= 0)
 					R = i - 1;
 				else
